# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order/{id}/close": {
            "patch": {
                "tags": [
                    "Order"
                ],
                "summary": "Закрытие заказа",
                "operationId": "00efb78ff7f1d2f89d07db31f34cfb6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "number": {
                                                    "type": "string",
                                                    "example": "Заказ-0101-0001"
                                                },
                                                "number_of_item": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "cost": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 251
                                                },
                                                "date_of_creation": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-01-01"
                                                },
                                                "closing_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-01-01"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "email has been requred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Запись не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}